pyproject.toml
[tool.poetry]
name = "soundpass"
version = "1.0.0"
description = ""
authors = ["Igor Gorelkov <igorelkov@keenetica.ru>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.9"
telegram = "^0.0.1"
telebot = "^0.0.5"
pyyaml = "^6.0.2"
colorlog = "^6.9.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
sqlalchemy-utils = "^0.41.2"
coloredlogs = "^15.0.1"
aiogram = "^3.17.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
